merged my branch "user_admin_1" into trunk.
This branch includes several bug fixes as well and the first little bit of user_admin functionality, which is espressly geared towards admins.
If you go to the admin panel and edit a user, you'll see that you can change all of their data and now adjust their account status.



git-svn-id: svn+ssh://svn.locusfoc.us/home/svn/ram/trunk@308 7bd17f4f-1217-0410-bbc5-bc4367116c30