merged branch ticket_30 with trunk. The ticket still remains open because I was unable to complete the task. Must bring in re-enforcements on this.
Major Changes
-----------------------------------------------
Fixed a bug in the asset controller refactor which broke the bulk uploader
Fixed a bug in the grail which made it vomit if HTML was passed into the message.
added a predicate method to the user object which affords statements like this

post :edit, :id=>@asset.id if current_user.can_edit? @asset




git-svn-id: svn+ssh://svn.locusfoc.us/home/svn/ram/trunk@486 7bd17f4f-1217-0410-bbc5-bc4367116c30