Merged ticket_111 with trunk.
Major Changes
-----------------------------------------------
Modified Asset.search to accept a hash or a string
The reason I had to do this is because user has_many :assets (but only ones, which he created not ones he has access to). 
All the assets are added by quentin and that is why he can add them to his briefcase but others could not.

I changed Asset.search to accept either a string or a hash, which has keys for "id","filename","description","name".

This way you can search for a specific asset with many attributes mapped to specific keys in the asset record.



git-svn-id: svn+ssh://svn.locusfoc.us/home/svn/ram/trunk@811 7bd17f4f-1217-0410-bbc5-bc4367116c30